backend/pyproject.toml
[project]
name = "fintrack-backend"
version = "0.1.0"
description = "Backend service for multi-currency finance tracker"
requires-python = ">=3.10"
authors = ["Finance Team <dev@example.com>"]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg[binary]",
    "pydantic>=2.5",
    "pydantic-settings",
    "email-validator",
    "passlib[bcrypt]",
    "bcrypt==3.2.2",
    "pyjwt[crypto]",
    "python-multipart",
    "httpx",
    "apscheduler",
    "python-dateutil"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "faker"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
